[vulkan_layer] Store shader hashes on pipeline creation
After pipeline creation the shader might be freed and the handle re-used
for different shaders. These seems to cause a lot of problems, as newer
shaders override older ones in the hash table. Therefore we cannot rely
on the shaders hash after the shader module was destroyed and need to store
it before that can happen.